requirements.txt
discord.py
python-dotenv
satisfactory-api-client
